Add Sequrity and improved filter

+++ src/main/java/com/bolilyivs/cookbook/Controller/RecipeController.java
@@ -8,7 +8,9 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -49,11 +51,26 @@ public class RecipeController {
     public List<Recipe> find(@RequestBody RecipeFinder finder){
         System.out.println(finder);
         List<Recipe> recipeList;
+
         finder.tags.add("all");
         finder.ingredients.add("all");
+
+        Direction dir = Direction.DESC;
+        if(finder.getSortingDir().equals("asc"))
+            dir = Direction.ASC;
+
+        String sorting = "rating";
+        if(finder.getSorting().equals("title"))
+            sorting = "title";
+        if(finder.getSorting().equals("last"))
+            sorting = "id";
+
+        System.out.println(sorting + " " + dir);
+
         recipeList = recipeRepo.searchAllTitleUsernameTagsIngredients(finder.title,finder.username,
                 finder.tags,(long) finder.tags.size(), finder.ingredients, (long) finder.ingredients.size(),
-                PageRequest.of(finder.page, finder.size, Direction.DESC, "rating"));
+                PageRequest.of(finder.page, finder.size, dir, sorting));
+
         recipeList.forEach((recipe -> {
             recipe.getTags().remove(new Tag("all"));
             recipe.getIngredients().remove(new Ingredient("all", "all"));
@@ -64,12 +81,10 @@ public class RecipeController {
 
     @PostMapping(path = "/find/count")
     public Long count(@RequestBody RecipeFinder finder){
-        System.out.println(finder);
         finder.tags.add("all");
         finder.ingredients.add("all");
         long count = recipeRepo.countTitleUsernameTagsIngredients(finder.title,finder.username,
                 finder.tags,(long) finder.tags.size(), finder.ingredients, (long) finder.ingredients.size());
-        System.out.println(count);
         return count;
     }
 
@@ -94,15 +109,24 @@ public class RecipeController {
 
     @PutMapping("/update/{id}")
     public Recipe update(@PathVariable("id") Recipe recipeDB, @RequestBody Recipe recipe){
-        recipeRepo.save(recipeDB);
-        recipeDB.setTitle(recipe.getTitle());
-        recipeDB.setDescription(recipe.getDescription());
-        recipeDB.setIngredients(recipe.getIngredients());
-        recipeDB.setTags(recipe.getTags());
-        recipeDB.getIngredients().add(new Ingredient("all", "all"));
-        recipeDB.getTags().add(new Tag("all"));
-        System.out.println(recipeDB);
-        return recipeRepo.save(recipeDB);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        System.out.println(auth.getAuthorities());
+        if(recipeDB.getAccount().getUsername().equals(auth.getName()) ||
+                auth.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN"))
+        ) {
+
+            recipeDB.setTitle(recipe.getTitle());
+            recipeDB.setDescription(recipe.getDescription());
+            recipeDB.setIngredients(recipe.getIngredients());
+            recipeDB.setTags(recipe.getTags());
+            recipeDB.getIngredients().add(new Ingredient("all", "all"));
+            recipeDB.getTags().add(new Tag("all"));
+            System.out.println(recipeDB);
+            return recipeRepo.save(recipeDB);
+        }else{
+            throw new UsernameNotFoundException("Bad");
+        }
+
     }
 
     @GetMapping("/rating/plus/{id}")
@@ -163,6 +187,13 @@ public class RecipeController {
 
     @DeleteMapping("delete/{id}")
     public void delete(@PathVariable("id") Recipe recipe){
-        recipeRepo.delete(recipe);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if(recipe.getAccount().getUsername().equals(auth.getName()) ||
+                auth.getAuthorities().contains(auth.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN")))
+        ){
+            recipeRepo.delete(recipe);
+        }else{
+            throw new UsernameNotFoundException("Bad");
+        }
     }
 }
